Split the round driver out of the learner module
direct_lstar.py held two things: the learner, and one particular way of driving
it in rounds.  The second is now fnr_synthesis.py -- the FNR gate, its pool
management, and the loop bookkeeping (_Best, _StallDetector, _discover,
_estimate_accuracy, _default_patience).

The learner knows nothing about rounds, families being resampled, or boundary
strings being fed back; it learns one hypothesis from one family.  Keeping the
driver alongside it made that relationship look like part of the algorithm.

direct_lstar 757 -> 494 lines, and its docstring now describes what the module
actually contains -- the probe loop, and the four objects it works through --
rather than the tree representation it stopped owning when MidfixTree landed.

modulo byte-identical: queries=747312, batches=890, fp@1024=1181, states=9.
